components/processing/libs/processor-xlsx/pyproject.toml
[project]
name = "processor-xlsx"
version = "0.0.1.dev"
dependencies = [
  "pyexcel==0.6.7",
  "pyexcel-xlsx==0.6.0",
  "openpyxl==3.0.10",
  "pyexcel-text==0.2.7.1",
  "faker",
  "fpdf2",
  "python-markdown-generator",
  "processor-base @ ${PROJECT_ROOT}/components/processing/libs/processor-base",
]
